requirements.txt
python-telegram-bot==21.8
anthropic==0.39.0
python-dotenv==1.0.0
